+++ src/mx_bluesky/hyperion/experiment_plans/__init__.py
@@ -3,11 +3,11 @@
 The __all__ list in here are the plans that are externally available from outside Hyperion.
 """
 
-from mx_bluesky.hyperion.experiment_plans.flyscan_xray_centre_plan import (
-    flyscan_xray_centre,
-)
 from mx_bluesky.hyperion.experiment_plans.grid_detect_then_xray_centre_plan import (
     grid_detect_then_xray_centre,
+)
+from mx_bluesky.hyperion.experiment_plans.hyperion_flyscan_xray_centre_plan import (
+    hyperion_flyscan_xray_centre,
 )
 from mx_bluesky.hyperion.experiment_plans.load_centre_collect_full_plan import (
     load_centre_collect_full,
@@ -24,7 +24,7 @@ from mx_bluesky.hyperion.experiment_plans.rotation_scan_plan import (
 )
 
 __all__ = [
-    "flyscan_xray_centre",
+    "hyperion_flyscan_xray_centre",
     "grid_detect_then_xray_centre",
     "rotation_scan",
     "pin_tip_centre_then_xray_centre",
