+++ src/mx_bluesky/hyperion/experiment_plans/grid_detect_then_xray_centre_plan.py
@@ -2,34 +2,14 @@ from __future__ import annotations
 
 from pathlib import Path
 
-import pydantic
 from blueapi.core import BlueskyContext
 from bluesky import plan_stubs as bps
 from bluesky import preprocessors as bpp
 from bluesky.preprocessors import subs_decorator
 from bluesky.utils import MsgGenerator
-from dodal.devices.aperturescatterguard import ApertureScatterguard
-from dodal.devices.attenuator.attenuator import BinaryFilterAttenuator
-from dodal.devices.backlight import Backlight, BacklightPosition
-from dodal.devices.dcm import DCM
-from dodal.devices.detector.detector_motion import DetectorMotion
+from dodal.devices.backlight import BacklightPosition
 from dodal.devices.eiger import EigerDetector
-from dodal.devices.fast_grid_scan import PandAFastGridScan, ZebraFastGridScan
-from dodal.devices.flux import Flux
-from dodal.devices.i03.beamstop import Beamstop
-from dodal.devices.oav.oav_detector import OAV
 from dodal.devices.oav.oav_parameters import OAVParameters
-from dodal.devices.oav.pin_image_recognition import PinTipDetection
-from dodal.devices.robot import BartRobot
-from dodal.devices.s4_slit_gaps import S4SlitGaps
-from dodal.devices.smargon import Smargon
-from dodal.devices.synchrotron import Synchrotron
-from dodal.devices.undulator import Undulator
-from dodal.devices.xbpm_feedback import XBPMFeedback
-from dodal.devices.zebra.zebra import Zebra
-from dodal.devices.zebra.zebra_controlled_shutter import ZebraShutter
-from dodal.devices.zocalo import ZocaloResults
-from ophyd_async.fastcs.panda import HDFPanda
 
 from mx_bluesky.common.external_interaction.callbacks.common.grid_detection_callback import (
     GridDetectionCallback,
@@ -39,9 +19,14 @@ from mx_bluesky.common.external_interaction.callbacks.xray_centre.ispyb_callback
     ispyb_activation_wrapper,
 )
 from mx_bluesky.common.parameters.constants import OavConstants
+from mx_bluesky.common.plans.common_flyscan_xray_centre_plan import (
+    flyscan_xray_centre_no_move,
+)
 from mx_bluesky.common.utils.context import device_composite_from_context
 from mx_bluesky.common.utils.log import LOGGER
-from mx_bluesky.common.xrc_result import XRayCentreEventHandler
+from mx_bluesky.common.xrc_result import (
+    XRayCentreEventHandler,
+)
 from mx_bluesky.hyperion.device_setup_plans.manipulate_sample import (
     move_aperture_if_required,
 )
@@ -51,8 +36,8 @@ from mx_bluesky.hyperion.device_setup_plans.utils import (
 from mx_bluesky.hyperion.experiment_plans.change_aperture_then_move_plan import (
     change_aperture_then_move_to_xtal,
 )
-from mx_bluesky.hyperion.experiment_plans.flyscan_xray_centre_plan import (
-    flyscan_xray_centre_no_move,
+from mx_bluesky.hyperion.experiment_plans.hyperion_flyscan_xray_centre_plan import (
+    construct_hyperion_specific_features,
 )
 from mx_bluesky.hyperion.experiment_plans.oav_grid_detection_plan import (
     OavGridDetectionComposite,
@@ -60,40 +45,13 @@ from mx_bluesky.hyperion.experiment_plans.oav_grid_detection_plan import (
 )
 from mx_bluesky.hyperion.parameters.constants import CONST
 from mx_bluesky.hyperion.parameters.device_composites import (
+    GridDetectThenXRayCentreComposite,
     HyperionFlyScanXRayCentreComposite,
 )
 from mx_bluesky.hyperion.parameters.gridscan import (
     GridScanWithEdgeDetect,
     HyperionSpecifiedThreeDGridScan,
 )
-
-
-@pydantic.dataclasses.dataclass(config={"arbitrary_types_allowed": True})
-class GridDetectThenXRayCentreComposite:
-    """All devices which are directly or indirectly required by this plan"""
-
-    aperture_scatterguard: ApertureScatterguard
-    attenuator: BinaryFilterAttenuator
-    backlight: Backlight
-    beamstop: Beamstop
-    dcm: DCM
-    detector_motion: DetectorMotion
-    eiger: EigerDetector
-    zebra_fast_grid_scan: ZebraFastGridScan
-    flux: Flux
-    oav: OAV
-    pin_tip_detection: PinTipDetection
-    smargon: Smargon
-    synchrotron: Synchrotron
-    s4_slit_gaps: S4SlitGaps
-    undulator: Undulator
-    xbpm_feedback: XBPMFeedback
-    zebra: Zebra
-    zocalo: ZocaloResults
-    panda: HDFPanda
-    panda_fast_grid_scan: PandAFastGridScan
-    robot: BartRobot
-    sample_shutter: ZebraShutter
 
 
 def create_devices(context: BlueskyContext) -> GridDetectThenXRayCentreComposite:
@@ -142,14 +100,6 @@ def detect_grid_and_do_gridscan(
             parameters.box_size_um,
         )
 
-    if parameters.selected_aperture:
-        # Start moving the aperture/scatterguard into position without moving it in
-        yield from bps.prepare(
-            composite.aperture_scatterguard,
-            parameters.selected_aperture,
-            group=CONST.WAIT.GRID_READY_FOR_DC,
-        )
-
     yield from run_grid_detection_plan(
         oav_params,
         snapshot_template,
@@ -166,31 +116,34 @@ def detect_grid_and_do_gridscan(
         group=CONST.WAIT.GRID_READY_FOR_DC,
     )
 
-    yield from flyscan_xray_centre_no_move(
-        HyperionFlyScanXRayCentreComposite(
-            aperture_scatterguard=composite.aperture_scatterguard,
-            attenuator=composite.attenuator,
-            backlight=composite.backlight,
-            eiger=composite.eiger,
-            panda_fast_grid_scan=composite.panda_fast_grid_scan,
-            flux=composite.flux,
-            s4_slit_gaps=composite.s4_slit_gaps,
-            smargon=composite.smargon,
-            undulator=composite.undulator,
-            synchrotron=composite.synchrotron,
-            xbpm_feedback=composite.xbpm_feedback,
-            zebra=composite.zebra,
-            zocalo=composite.zocalo,
-            panda=composite.panda,
-            zebra_fast_grid_scan=composite.zebra_fast_grid_scan,
-            dcm=composite.dcm,
-            robot=composite.robot,
-            sample_shutter=composite.sample_shutter,
-        ),
-        create_parameters_for_flyscan_xray_centre(
-            parameters, grid_params_callback.get_grid_parameters()
-        ),
+    xrc_composite = HyperionFlyScanXRayCentreComposite(
+        aperture_scatterguard=composite.aperture_scatterguard,
+        attenuator=composite.attenuator,
+        backlight=composite.backlight,
+        eiger=composite.eiger,
+        panda_fast_grid_scan=composite.panda_fast_grid_scan,
+        flux=composite.flux,
+        s4_slit_gaps=composite.s4_slit_gaps,
+        smargon=composite.smargon,
+        undulator=composite.undulator,
+        synchrotron=composite.synchrotron,
+        xbpm_feedback=composite.xbpm_feedback,
+        zebra=composite.zebra,
+        zocalo=composite.zocalo,
+        panda=composite.panda,
+        zebra_fast_grid_scan=composite.zebra_fast_grid_scan,
+        dcm=composite.dcm,
+        robot=composite.robot,
+        sample_shutter=composite.sample_shutter,
     )
+
+    params = create_parameters_for_flyscan_xray_centre(
+        parameters, grid_params_callback.get_grid_parameters()
+    )
+
+    feature_controlled = construct_hyperion_specific_features(xrc_composite, params)
+
+    yield from flyscan_xray_centre_no_move(xrc_composite, params, feature_controlled)
 
 
 def grid_detect_then_xray_centre(
