+++ tests/system_tests/hyperion/experiment_plans/test_fgs_plan.py
@@ -5,6 +5,7 @@ from unittest.mock import MagicMock, patch
 import bluesky.plan_stubs as bps
 import bluesky.preprocessors as bpp
 import pytest
+import pytest_asyncio
 from bluesky.run_engine import RunEngine
 from dodal.beamlines import i03
 from dodal.devices.aperturescatterguard import ApertureValue
@@ -12,6 +13,12 @@ from dodal.devices.smargon import Smargon
 from ophyd.sim import NullStatus
 from ophyd_async.testing import set_mock_value
 
+from mx_bluesky.common.device_setup_plans.xbpm_feedback import (
+    transmission_and_xbpm_feedback_for_collection_decorator,
+)
+from mx_bluesky.common.external_interaction.callbacks.common.callback_util import (
+    create_gridscan_callbacks,
+)
 from mx_bluesky.common.external_interaction.callbacks.xray_centre.ispyb_callback import (
     GridscanISPyBCallback,
 )
@@ -19,24 +26,16 @@ from mx_bluesky.common.external_interaction.callbacks.xray_centre.nexus_callback
     GridscanNexusFileCallback,
 )
 from mx_bluesky.common.external_interaction.ispyb.ispyb_store import IspybIds
-from mx_bluesky.common.plans.read_hardware import (
-    standard_read_hardware_during_collection,
-    standard_read_hardware_pre_collection,
-)
 from mx_bluesky.common.utils.exceptions import WarningException
-from mx_bluesky.hyperion.device_setup_plans.xbpm_feedback import (
-    transmission_and_xbpm_feedback_for_collection_decorator,
+from mx_bluesky.hyperion.device_setup_plans.read_hardware_for_setup import (
+    read_hardware_during_collection,
+    read_hardware_pre_collection,
 )
 from mx_bluesky.hyperion.experiment_plans.flyscan_xray_centre_plan import (
-    flyscan_xray_centre,
-)
-from mx_bluesky.hyperion.external_interaction.callbacks.__main__ import (
-    create_gridscan_callbacks,
+    HyperionFlyScanXRayCentreComposite,
+    hyperion_flyscan_xray_centre,
 )
 from mx_bluesky.hyperion.parameters.constants import CONST
-from mx_bluesky.hyperion.parameters.device_composites import (
-    HyperionFlyScanXRayCentreComposite,
-)
 from mx_bluesky.hyperion.parameters.gridscan import HyperionSpecifiedThreeDGridScan
 from tests.conftest import default_raw_gridscan_params
 
@@ -67,7 +66,7 @@ def reset_positions(smargon: Smargon):
     yield from bps.mv(smargon.x, -1, smargon.y, -1, smargon.z, -1)  # type: ignore # See: https://github.com/bluesky/bluesky/issues/1809
 
 
-@pytest.fixture
+@pytest_asyncio.fixture
 async def fxc_composite():
     with (
         patch("dodal.devices.zocalo.zocalo_results._get_zocalo_connection"),
@@ -77,12 +76,10 @@ async def fxc_composite():
         zocalo = i03.zocalo()
 
     composite = HyperionFlyScanXRayCentreComposite(
-        attenuator=i03.attenuator(connect_immediately=True, mock=True),
-        aperture_scatterguard=i03.aperture_scatterguard(
-            connect_immediately=True, mock=True
-        ),
-        backlight=i03.backlight(mock=True),
-        dcm=i03.dcm(fake_with_ophyd_sim=True),
+        attenuator=i03.attenuator(),
+        aperture_scatterguard=i03.aperture_scatterguard(),
+        backlight=i03.backlight(),
+        dcm=i03.dcm(connect_immediately=True, mock=True),
         eiger=i03.eiger(),
         zebra_fast_grid_scan=i03.zebra_fast_grid_scan(),
         flux=i03.flux(connect_immediately=True, mock=True),
@@ -130,10 +127,10 @@ def test_read_hardware_pre_collection(
 ):
     @bpp.run_decorator()
     def read_run(u, s, g, r, a, f, dcm, ap_sg, sm):
-        yield from standard_read_hardware_pre_collection(
+        yield from read_hardware_pre_collection(
             undulator=u, synchrotron=s, s4_slit_gaps=g, dcm=dcm, smargon=sm
         )
-        yield from standard_read_hardware_during_collection(
+        yield from read_hardware_during_collection(
             ap_sg, a, f, dcm, fxc_composite.eiger
         )
 
@@ -164,10 +161,8 @@ async def test_xbpm_feedback_decorator(
     # in S03
 
     @transmission_and_xbpm_feedback_for_collection_decorator(
-        fxc_composite.undulator,
         fxc_composite.xbpm_feedback,
         fxc_composite.attenuator,
-        fxc_composite.dcm,
         params.transmission_frac,
     )
     def decorated_plan():
@@ -182,11 +177,11 @@ async def test_xbpm_feedback_decorator(
 @patch("bluesky.plan_stubs.kickoff", autospec=True)
 @patch("bluesky.plan_stubs.complete", autospec=True)
 @patch(
-    "mx_bluesky.hyperion.experiment_plans.flyscan_xray_centre_plan.run_gridscan_and_move",
+    "mx_bluesky.hyperion.experiment_plans.hyperion_flyscan_xray_centre_plan.run_gridscan_and_move",
     autospec=True,
 )
 @patch(
-    "mx_bluesky.hyperion.experiment_plans.flyscan_xray_centre_plan.tidy_up_zebra_after_gridscan",
+    "mx_bluesky.hyperion.experiment_plans.hyperion_flyscan_xray_centre_plan.tidy_up_zebra_after_gridscan",
     autospec=True,
 )
 def test_full_plan_tidies_at_end(
@@ -208,7 +203,7 @@ def test_full_plan_tidies_at_end(
         data_collection_ids=(0, 0), data_collection_group_id=0, grid_ids=(0,)
     )
     [RE.subscribe(cb) for cb in callbacks]
-    RE(flyscan_xray_centre(fxc_composite, params))
+    RE(hyperion_flyscan_xray_centre(fxc_composite, params))
     set_shutter_to_manual.assert_called_once()
 
 
@@ -217,11 +212,11 @@ def test_full_plan_tidies_at_end(
 @patch("bluesky.plan_stubs.kickoff", autospec=True)
 @patch("bluesky.plan_stubs.complete", autospec=True)
 @patch(
-    "mx_bluesky.hyperion.experiment_plans.flyscan_xray_centre_plan.run_gridscan_and_move",
+    "mx_bluesky.hyperion.experiment_plans.hyperion_flyscan_xray_centre_plan.run_gridscan_and_move",
     autospec=True,
 )
 @patch(
-    "mx_bluesky.hyperion.experiment_plans.flyscan_xray_centre_plan.tidy_up_zebra_after_gridscan",
+    "mx_bluesky.hyperion.experiment_plans.hyperion_flyscan_xray_centre_plan.tidy_up_zebra_after_gridscan",
     autospec=True,
 )
 def test_full_plan_tidies_at_end_when_plan_fails(
@@ -238,7 +233,7 @@ def test_full_plan_tidies_at_end_when_plan_fails(
 
     run_gridscan_and_move.side_effect = _Exception()
     with pytest.raises(_Exception):
-        RE(flyscan_xray_centre(fxc_composite, params))
+        RE(hyperion_flyscan_xray_centre(fxc_composite, params))
     set_shutter_to_manual.assert_called_once()
 
 
@@ -264,7 +259,7 @@ def test_GIVEN_scan_invalid_WHEN_plan_run_THEN_ispyb_entry_made_but_no_zocalo_en
 
     [RE.subscribe(cb) for cb in callbacks]
     with pytest.raises(WarningException):
-        RE(flyscan_xray_centre(fxc_composite, params))
+        RE(hyperion_flyscan_xray_centre(fxc_composite, params))
 
     ids = ispyb_cb.ispyb_ids
     assert ids.data_collection_group_id is not None
@@ -302,7 +297,7 @@ async def test_complete_xray_centre_plan_with_no_callbacks_falls_back_to_centre(
 
     # [RE.subscribe(cb) for cb in callbacks]
     fxc_composite.zocalo.trigger = MagicMock(side_effect=zocalo_trigger)
-    RE(flyscan_xray_centre(fxc_composite, params))
+    RE(hyperion_flyscan_xray_centre(fxc_composite, params))
 
     # The following numbers are derived from the centre returned in fake_zocalo
     assert await fxc_composite.smargon.x.user_readback.get_value() == pytest.approx(-1)
@@ -331,7 +326,7 @@ async def test_complete_xray_centre_plan_with_callbacks_moves_to_centre(
     RE(reset_positions(fxc_composite.smargon))
 
     [RE.subscribe(cb) for cb in callbacks]
-    RE(flyscan_xray_centre(fxc_composite, params))
+    RE(hyperion_flyscan_xray_centre(fxc_composite, params))
 
     # The following numbers are derived from the centre returned in fake_zocalo
     assert await fxc_composite.smargon.x.user_readback.get_value() == pytest.approx(
