+++ tests/unit_tests/hyperion/device_setup_plans/test_xbpm_feedback.py
@@ -8,7 +8,7 @@ from dodal.devices.xbpm_feedback import Pause
 from ophyd.status import Status
 from ophyd_async.testing import set_mock_value
 
-from mx_bluesky.hyperion.device_setup_plans.xbpm_feedback import (
+from mx_bluesky.common.device_setup_plans.xbpm_feedback import (
     transmission_and_xbpm_feedback_for_collection_decorator,
 )
