+++ tests/unit_tests/hyperion/experiment_plans/test_robot_load_then_centre.py
@@ -8,10 +8,10 @@ from bluesky.utils import Msg
 from dodal.devices.i03.beamstop import BeamstopPositions
 from dodal.devices.robot import SampleLocation
 
-from mx_bluesky.hyperion.device_setup_plans.check_beamstop import BeamstopException
-from mx_bluesky.hyperion.experiment_plans.flyscan_xray_centre_plan import (
+from mx_bluesky.common.plans.common_flyscan_xray_centre_plan import (
     _fire_xray_centre_result_event,
 )
+from mx_bluesky.hyperion.device_setup_plans.check_beamstop import BeamstopException
 from mx_bluesky.hyperion.experiment_plans.grid_detect_then_xray_centre_plan import (
     GridDetectThenXRayCentreComposite,
 )
